TST: load()

+++ tests/test_load.py
@@ -1,33 +1,74 @@
 import os
-import pytest
-import tempfile
 import shutil
 
-import audmodel
+import pytest
+
 import audeer
-import audfactory
+import audmodel
 
 
-@pytest.mark.usefixtures('create')
-@pytest.mark.parametrize('verbose', [True, False])
-def test_load(verbose):
-    uid = pytest.UIDS[0]
-    root = os.path.join(tempfile._get_default_tempdir(), 'audmodel')
-    model_root = audmodel.load(uid, root=root, verbose=verbose)
-    expected_uid = os.path.basename(model_root)
-    assert expected_uid == uid
-    group_id = f'{audmodel.core.define.defaults.GROUP_ID}.{pytest.SUBGROUP}'
-    assert model_root == os.path.join(
-        root,
-        pytest.REPOSITORY,
-        audfactory.group_id_to_path(group_id),
-        pytest.NAME,
-        pytest.VERSION,
-        uid,
-    )
-    x = [os.path.basename(file) for file in
-         audeer.list_file_names(pytest.ROOT)]
-    y = [os.path.basename(file) for file in
-         audeer.list_file_names(model_root)]
-    assert x == y
-    shutil.rmtree(root)
+audmodel.config.BACKEND_HOST = pytest.BACKEND_HOST
+audmodel.config.CACHE_ROOT = pytest.CACHE_ROOT
+
+MODEL_FILES = {
+    '1.0.0': ['test', 'sub/test'],
+    '2.0.0': ['other', 'sub/test'],
+}
+SUBGROUP = f'{pytest.ID}.publish'
+VERSION = '1.0.0'
+
+MODEL_ID = audmodel.uid(pytest.NAME, pytest.PARAMS, subgroup=SUBGROUP)
+
+
+def clear_root(root: str):
+    root = audeer.safe_path(root)
+    if os.path.exists(root):
+        shutil.rmtree(root)
+
+
+@pytest.fixture(
+    scope='module',
+    autouse=True,
+)
+def fixture_publish_model():
+
+    for version, files in MODEL_FILES.items():
+
+        clear_root(pytest.MODEL_ROOT)
+
+        for file in files:
+            path = os.path.join(pytest.MODEL_ROOT, file)
+            audeer.mkdir(os.path.dirname(path))
+            with open(path, 'w'):
+                pass
+
+        audmodel.publish(
+            pytest.MODEL_ROOT,
+            pytest.NAME,
+            pytest.PARAMS,
+            version,
+            subgroup=SUBGROUP,
+        )
+
+    yield
+
+    clear_root(pytest.MODEL_ROOT)
+
+
+@pytest.mark.parametrize(
+    'uid, version',
+    (
+        (MODEL_ID, '1.0.0'),
+        (MODEL_ID, '2.0.0'),
+        (MODEL_ID, None),
+    ),
+)
+def test_load(uid, version):
+
+    root = audmodel.load(uid, version=version)
+
+    if version is None:
+        version = audmodel.latest_version(uid)
+
+    files = audmodel.core.api.scan_files(root)
+    assert sorted(MODEL_FILES[version]) == sorted(files)
